Added changes after review
